requirements.txt
absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==4.1
django-admin-env-notice==0.4
django-admin-honeypot==1.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-health-check==3.17.0
django-material-admin==1.8.6
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
flatbuffers==22.9.24
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.9.10
six==1.16.0
sqlparse==0.4.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
